+++ src/main/webclient/src/client/app-wrapper.tsx
@@ -1,25 +1,37 @@
 import { ReactNode } from "react";
 import { BrowserRouter } from "react-router-dom";
 import { MantineProvider, createTheme } from "@mantine/core";
-import { ChakraProvider, extendBaseTheme } from "@chakra-ui/react";
+import { ChakraProvider, extendTheme } from "@chakra-ui/react";
+import { appProps } from "../config/config";
 
-const mantineTheme = createTheme({
-  /** Put your mantine theme override here */
-});
 
-const chakraTheme = extendBaseTheme({
-  components: {
-    //default themes for components that are using
-    //ChakraBaseProvider helps to reduse initial JS payload is the size of the component themes
-  },
-});
+export const AppWrapper = ({children}: {children: ReactNode}) => {
+  const mantineTheme = createTheme({
+    /** Put your mantine theme override here */
+  });
 
-export const AppWrapper = ({ children }: { children: ReactNode }) => {
+  const chakraTheme = extendTheme({
+    colors: {
+      primary: {
+        500: appProps.appMainProps.ui.colors.primary,
+      },
+      secondary: {
+        500: appProps.appMainProps.ui.colors.secondary,
+      },
+      background: {
+        500: appProps.appMainProps.ui.colors.background
+      },
+      accent: {
+        500: appProps.appMainProps.ui.colors.accent
+      }
+    },
+  });
+
   return (
     <BrowserRouter>
-        <MantineProvider theme={mantineTheme}>
-          <ChakraProvider>{children}</ChakraProvider>
-        </MantineProvider>
-      </BrowserRouter>
+      <MantineProvider theme={mantineTheme}>
+        <ChakraProvider theme={chakraTheme}>{children}</ChakraProvider>
+      </MantineProvider>
+    </BrowserRouter>
   );
 };
